requirements-dev.txt
asgiref==3.2.10
autopep8==1.5.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
django-bootstrap-form==3.4
django-heroku==0.3.1
gunicorn==20.0.4
mysqlclient==2.0.1
psycopg2==2.8.5
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
static3==0.7.0
toml==0.10.1
whitenoise==5.2.0
